- [mrpt-base] MRPT moves from ANN to FLANN for kd-tree library:   - icp-slam is now ~twice faster!!   - New namespace mrpt_flann, holding a modified version of FLANN.   - A modified FLANN class has been created, mrpt_flann::KDTreeSingleIndexAdaptor, to allow building a kd-tree index and performing search without duplicating all the dataset within the KD-tree index.   - mrpt::math::KDTreeCapable maintains its API, so the change is backward compatible   - New class mrpt::math::KDTreeMatrixAdaptor to build KD-tree directly from Eigen matrices.

+++ libs/slam/src/slam/CGridMapAligner.cpp
@@ -564,7 +564,7 @@ CPosePDFPtr CGridMapAligner::AlignPDF_robustMatch(
 
 						CPoint2D p2_j_local;
 						vector<float>  matches_x,matches_y, matches_dist;
-						std::vector<size_t> matches_idx;
+						std::vector<int> matches_idx;
 
 						CPoint2DPDFGaussian  pdf_M2_j;
 						CPoint2DPDFGaussian  pdf_M1_i;
